fix: Extract FakeEndpoint and Benchmark endpoints to separate service
