+++ hash_test.go
@@ -11,34 +11,6 @@ import (
 	"google.golang.org/grpc"
 )
 
-type mockedGRPCClient struct {
-	mock.Mock
-}
-
-func (m *mockedGRPCClient) Hash(ctx context.Context, in *protobuf.HashRequest, opts ...grpc.CallOption) (*protobuf.HashResponse, error) {
-	args := m.Called(ctx, in)
-
-	return args.Get(0).(*protobuf.HashResponse), args.Error(1)
-}
-
-func (m *mockedGRPCClient) Sign(ctx context.Context, in *protobuf.SignRequest, opts ...grpc.CallOption) (*protobuf.SignResponse, error) {
-	args := m.Called(ctx, in)
-
-	return args.Get(0).(*protobuf.SignResponse), args.Error(1)
-}
-
-func (m *mockedGRPCClient) Benchmark(ctx context.Context, in *protobuf.BenchmarkRequest, opts ...grpc.CallOption) (*protobuf.BenchmarkResponse, error) {
-	args := m.Called(ctx, in)
-
-	return args.Get(0).(*protobuf.BenchmarkResponse), args.Error(1)
-}
-
-func (m *mockedGRPCClient) FakeEndpoint(ctx context.Context, in *protobuf.FakeEndpointRequest, opts ...grpc.CallOption) (*protobuf.FakeEndpointResponse, error) {
-	args := m.Called(ctx, in)
-
-	return args.Get(0).(*protobuf.FakeEndpointResponse), args.Error(1)
-}
-
 func TestLibrary_HashData(t *testing.T) {
 	mockedClient := &mockedGRPCClient{}
